+++ app/login/page.tsx
@@ -5,7 +5,7 @@ import { TextField, Container, Box, Button, Typography } from '@mui/material';
 
 import NavigationLink from '../../components/atoms/NavigationLink';
 
-import useLogin from '../../hooks/useLogin';
+import useLogin from '../../hooks/app/useLogin';
 
 import { mainTheme } from '../../styles/theme'
 import ThemeWrapper from '../../styles/ThemeWrapper';
