+++ components/organisms/ClientSideComponent.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useEffect, useState } from 'react';
-import usePwaStatus from '../../hooks/usePwaStatus';
+import usePwaStatus from '../../hooks/app/usePwaStatus';
 import NotPwaHeader from './notPwaHeader';
 
 const ClientSideComponent: React.FC<{ children: React.ReactNode }> = ({ children }) => {
